Cast int64 literals to double in jsonutil_is_int64 tests (#99)
Newer clang (14+, including current Apple clang) enables
-Wimplicit-const-int-float-conversion under -Wall, which combined with
the project's -Werror breaks ./build.sh --unit on these two assertions:

    EXPECT_TRUE(jsonutil_is_int64(INT64_MAX >> 1));
    EXPECT_TRUE(jsonutil_is_int64(8223372036854775807LL));

jsonutil_is_int64 takes a double, so passing a long long literal
implicitly narrows. The narrowed values still round-trip cleanly
through the function (it does double -> int64 -> double equality),
so the assertions remain TRUE; the cast just acknowledges the
narrowing the signature already implies.

(cherry picked from commit 66101ce33fff7d3c0c64a4b221ff36875edffccf)

+++ tst/unit/util_test.cc
@@ -58,8 +58,8 @@ TEST_F(UtilTest, testIsInt64) {
     EXPECT_TRUE(jsonutil_is_int64(INT16_MIN));
     EXPECT_TRUE(jsonutil_is_int64(INT32_MAX));
     EXPECT_TRUE(jsonutil_is_int64(INT32_MIN));
-    EXPECT_TRUE(jsonutil_is_int64(INT64_MAX >> 1));
-    EXPECT_TRUE(jsonutil_is_int64(8223372036854775807LL));
+    EXPECT_TRUE(jsonutil_is_int64(static_cast<double>(INT64_MAX >> 1)));
+    EXPECT_TRUE(jsonutil_is_int64(static_cast<double>(8223372036854775807LL)));
     EXPECT_TRUE(jsonutil_is_int64(INT64_MIN));
     EXPECT_FALSE(jsonutil_is_int64(1e28));      // out of range of int64
     EXPECT_FALSE(jsonutil_is_int64(1.7e308));   // out of range of int64
